Refactored float to decimal for financial data such as amount.

+++ api/interfaces/account_balance.py
@@ -1,8 +1,10 @@
 #!/usr/bin/env python3
+from decimal import Decimal
+
 from pydantic.dataclasses import dataclass
 
 
 @dataclass
 class AccountBalance:
     currency: str
-    available_balance: float
+    available_balance: Decimal
